Updated Dashboard Page
Initial update of dashboard table and graphs. Added determination of initial temperature units, and converts to user-selected units. Calculates VPDleaf, VPDair, and LAD. Shows max, min, and average values for each, as well as start and end time of data collection, and time interval between data points. Created button for the user to download a pdf of the report.